Update to support bazel and buck tests (including plugins)

+++ jenkins/gerrit-buck-build-plugin-manual.sh
@@ -7,25 +7,28 @@
 # out the proper version of Gerrit, removing the plugin if it
 # exists, then use git read-tree to put the plugin we're
 # building in place.
-git checkout gerrit/{branch}
-rm -rf plugins/{name}
-git fetch https://gerrit.googlesource.com/plugins/{name} $REFS_CHANGE
-git read-tree -u --prefix=plugins/{name} FETCH_HEAD
+if [ -f 'BUCK' ]
+then
+  git checkout gerrit/{branch}
+  rm -rf plugins/{name}
+  git fetch https://gerrit.googlesource.com/plugins/{name} $REFS_CHANGE
+  git read-tree -u --prefix=plugins/{name} FETCH_HEAD
 
-TARGETS=$(echo "{targets}" | sed -e 's/{{name}}/{name}/g')
+  TARGETS=$(echo "{targets}" | sed -e 's/{{name}}/{name}/g')
 
-SOURCE_LEVEL=$(grep "source_level" .buckconfig || echo "source_level=7")
-. set-java.sh $(echo $SOURCE_LEVEL | cut -d '=' -f 2 | tr -d '[[:space:]]')
+  SOURCE_LEVEL=$(grep "source_level" .buckconfig || echo "source_level=7")
+  . set-java.sh $(echo $SOURCE_LEVEL | cut -d '=' -f 2 | tr -d '[[:space:]]')
 
-buck build -v 3 $TARGETS
+  buck build -v 3 $TARGETS
 
-for JAR in $(buck targets --show_output $TARGETS | awk '{{print $2}}')
-do
-    PLUGIN_VERSION=$(git describe  --always origin/{branch})
-    echo -e "Implementation-Version: $PLUGIN_VERSION" > MANIFEST.MF
-    jar ufm $JAR MANIFEST.MF && rm MANIFEST.MF
+  for JAR in $(buck targets --show_output $TARGETS | awk '{{print $2}}')
+  do
+      PLUGIN_VERSION=$(git describe  --always origin/{branch})
+      echo -e "Implementation-Version: $PLUGIN_VERSION" > MANIFEST.MF
+      jar ufm $JAR MANIFEST.MF && rm MANIFEST.MF
 
-    DEST_JAR=buck-out/gen/plugins/{name}/$(basename $JAR)
-    [ "$JAR" -ef "$DEST_JAR" ] || mv $JAR $DEST_JAR
-    echo "$PLUGIN_VERSION" > buck-out/gen/plugins/{name}/$(basename $JAR-version)
-done
+      DEST_JAR=buck-out/gen/plugins/{name}/$(basename $JAR)
+      [ "$JAR" -ef "$DEST_JAR" ] || mv $JAR $DEST_JAR
+      echo "$PLUGIN_VERSION" > buck-out/gen/plugins/{name}/$(basename $JAR-version)
+  done
+fi
